adminExport: Run the survey file generation in a worker pool
fixes #309

This will run the export in a different thread, leaving the server free
to answer all other requests.

An eventual task to produce those files could also make use of this
pool.
