+++ packages/evolution-backend/package.json
@@ -33,7 +33,8 @@
         "moment": "^2.29.4",
         "papaparse": "^5.3.0",
         "typescript": "^4.9.4",
-        "uuid": "^8.3.2"
+        "uuid": "^8.3.2",
+        "workerpool": "^6.2.1"
     },
     "devDependencies": {
         "@types/jest": "^26.0.3",
